+++ nDead.c
@@ -183,7 +183,8 @@ static int nDeadDisplay(dspConfig_t *dsp)
         buf[0] = 0;
         /* We were unable to determine the cause of death. */
         utAppendShip(buf, snum);
-        sprintf(cbuf, "dead: %s was killed by %d(%d).", buf, kb, detail);
+        sprintf(cbuf, "dead: %s was killed by %d(%d).", buf,
+                (int)kb, (int)detail);
         utLog(cbuf);
 
         cprintf(8,0,ALIGN_CENTER,"#%d#%s%s",
